Treasure_Hunt
This program asks the user to input a direction (north east south or west). The program randomly generates a location of the treasure within a 4x4 matrix. Each time the user types in the name of a direction, they search in 1 position in said direction until they find the treasure.